Rename AgentInfoEncodeMeta field to affect breaking change

+++ crates/kitsune_p2p/bootstrap/benches/bench.rs
@@ -59,7 +59,7 @@ fn bootstrap(bench: &mut Criterion) {
                 let info = AgentInfoSigned::sign(
                     space.clone(),
                     Arc::new(fixt!(KitsuneAgent, Unpredictable)),
-                    u32::MAX / 4,
+                    u32::MAX as u64 / 4,
                     fixt!(UrlList, Empty),
                     0,
                     std::time::UNIX_EPOCH.elapsed().unwrap().as_millis() as u64 + 60_000_000,
